Extract connection protocol FSM
