+++ aries_vcx/src/handlers/connection/mod.rs
@@ -1,8 +1,4 @@
-pub mod pairwise_info;
 pub mod cloud_agent;
 pub mod legacy_agent_info;
 pub mod connection;
-pub mod invitee;
-pub mod inviter;
 pub mod public_agent;
-mod util;
