+++ aries_vcx/src/protocols/connection/invitee/states/requested.rs
@@ -1,5 +1,5 @@
-use crate::handlers::connection::invitee::states::initial::InitialState;
-use crate::handlers::connection::invitee::states::responded::RespondedState;
+use crate::protocols::connection::invitee::states::initial::InitialState;
+use crate::protocols::connection::invitee::states::responded::RespondedState;
 use crate::messages::connection::did_doc::DidDoc;
 use crate::messages::connection::problem_report::ProblemReport;
 use crate::messages::connection::request::Request;
