+++ aries_vcx/src/protocols/connection/inviter/states/complete.rs
@@ -1,20 +1,20 @@
-use std::future::Future;
 use std::clone::Clone;
+use std::future::Future;
 
 use crate::error::VcxResult;
-use crate::handlers::connection::util::handle_ping;
 use crate::messages::a2a::A2AMessage;
 use crate::messages::a2a::protocol_registry::ProtocolRegistry;
 use crate::messages::connection::did_doc::DidDoc;
 use crate::messages::discovery::disclose::{Disclose, ProtocolDescriptor};
 use crate::messages::discovery::query::Query;
 use crate::messages::trust_ping::ping::Ping;
+use crate::protocols::connection::util::handle_ping;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct CompleteState {
     pub did_doc: DidDoc,
     pub protocols: Option<Vec<ProtocolDescriptor>>,
-    pub thread_id: Option<String>
+    pub thread_id: Option<String>,
 }
 
 impl From<(CompleteState, Vec<ProtocolDescriptor>)> for CompleteState {
@@ -26,13 +26,13 @@ impl From<(CompleteState, Vec<ProtocolDescriptor>)> for CompleteState {
 
 impl CompleteState {
     pub async fn handle_send_ping<F, T>(&self,
-                            comment: Option<String>,
-                            pw_vk: &str,
-                            send_message: F,
+                                        comment: Option<String>,
+                                        pw_vk: &str,
+                                        send_message: F,
     ) -> VcxResult<()>
-    where
-        F: Fn(String, DidDoc, A2AMessage) -> T,
-        T: Future<Output=VcxResult<()>>
+        where
+            F: Fn(String, DidDoc, A2AMessage) -> T,
+            T: Future<Output=VcxResult<()>>
     {
         let ping =
             Ping::create()
@@ -44,26 +44,26 @@ impl CompleteState {
     }
 
     pub async fn handle_ping<F, T>(&self,
-                       ping: &Ping,
-                       pw_vk: &str,
-                       send_message: F
+                                   ping: &Ping,
+                                   pw_vk: &str,
+                                   send_message: F,
     ) -> VcxResult<()>
-    where
-        F: Fn(String, DidDoc, A2AMessage) -> T,
-        T: Future<Output=VcxResult<()>>
+        where
+            F: Fn(String, DidDoc, A2AMessage) -> T,
+            T: Future<Output=VcxResult<()>>
     {
         handle_ping(ping, pw_vk, &self.did_doc, send_message).await
     }
 
     pub async fn handle_discover_features<F, T>(&self,
-                                    query: Option<String>,
-                                    comment: Option<String>,
-                                    pw_vk: &str,
-                                    send_message: F
+                                                query: Option<String>,
+                                                comment: Option<String>,
+                                                pw_vk: &str,
+                                                send_message: F,
     ) -> VcxResult<()>
-    where
-        F: Fn(String, DidDoc, A2AMessage) -> T,
-        T: Future<Output=VcxResult<()>>
+        where
+            F: Fn(String, DidDoc, A2AMessage) -> T,
+            T: Future<Output=VcxResult<()>>
     {
         let query_ =
             Query::create()
@@ -74,13 +74,13 @@ impl CompleteState {
     }
 
     pub async fn handle_discovery_query<F, T>(&self,
-                                  query: Query,
-                                  pw_vk: &str,
-                                  send_message: F
+                                              query: Query,
+                                              pw_vk: &str,
+                                              send_message: F,
     ) -> VcxResult<()>
-    where
-        F: Fn(String, DidDoc, A2AMessage) -> T,
-        T: Future<Output=VcxResult<()>>
+        where
+            F: Fn(String, DidDoc, A2AMessage) -> T,
+            T: Future<Output=VcxResult<()>>
     {
         let protocols = ProtocolRegistry::init().get_protocols_for_query(query.query.as_ref().map(String::as_str));
