+++ aries_vcx/src/protocols/connection/inviter/states/invited.rs
@@ -1,5 +1,5 @@
-use crate::handlers::connection::inviter::states::initial::InitialState;
-use crate::handlers::connection::inviter::states::requested::RequestedState;
+use crate::protocols::connection::inviter::states::initial::InitialState;
+use crate::protocols::connection::inviter::states::requested::RequestedState;
 use crate::messages::connection::invite::Invitation;
 use crate::messages::connection::problem_report::ProblemReport;
 use crate::messages::connection::request::Request;
