+++ aries_vcx/src/protocols/connection/inviter/states/requested.rs
@@ -1,5 +1,5 @@
-use crate::handlers::connection::inviter::states::initial::InitialState;
-use crate::handlers::connection::inviter::states::responded::RespondedState;
+use crate::protocols::connection::inviter::states::initial::InitialState;
+use crate::protocols::connection::inviter::states::responded::RespondedState;
 use crate::messages::connection::did_doc::DidDoc;
 use crate::messages::connection::problem_report::ProblemReport;
 use crate::messages::connection::response::SignedResponse;
