+++ aries_vcx/src/protocols/connection/mod.rs
@@ -1,0 +1,4 @@
+pub mod invitee;
+pub mod inviter;
+pub mod pairwise_info;
+mod util;
