+++ aries_vcx/src/protocols/connection/util.rs
