+++ aries_vcx/src/protocols/mod.rs
@@ -1,0 +1,4 @@
+pub mod connection;
+// pub mod issuance;
+// pub mod proof_presentation;
+// pub mod out_of_band;
